fix(ci): python coverage shard line limit — split oversized workspace repository tests
